refactor: support 패키지 삭제
