+++ src/main/java/nextstep/exception/AlreadyReservedScheduleException.java
@@ -1,4 +1,4 @@
-package nextstep.support.exception;
+package nextstep.exception;
 
 public class AlreadyReservedScheduleException extends RuntimeException {
     public AlreadyReservedScheduleException() {
