+++ src/main/java/nextstep/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package nextstep.support.exception;
+package nextstep.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
