+++ src/main/java/nextstep/exception/NoReservationException.java
@@ -1,4 +1,4 @@
-package nextstep.support.exception;
+package nextstep.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
