+++ src/test/java/nextstep/reservation_waiting/ReservationWaitingServiceTest.java
@@ -5,8 +5,8 @@ import nextstep.member.MemberDao;
 import nextstep.reservation.Reservation;
 import nextstep.reservation.ReservationDao;
 import nextstep.schedule.Schedule;
-import nextstep.support.exception.DuplicateEntityException;
-import nextstep.support.exception.NotOwnReservationWaitingException;
+import nextstep.exception.DuplicateEntityException;
+import nextstep.exception.NotOwnReservationWaitingException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
